pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "discord-forum-slack"
version = "1.0.0"
description = "Discord 포럼 채널 새 글을 Slack으로 자동 전송하는 봇"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3.0",
    "PyYAML>=6.0",
    "requests>=2.31.0",
    "brotli>=1.1.0",
]

[project.scripts]
discord-forum-slack = "discord_forum_slack.__main__:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["discord_forum_slack"]
